+++ app/interactors/comments/destroy.rb
@@ -5,13 +5,7 @@ module Comments
     delegate :comment, to: :context
 
     def call
-      context.fail!(error: "Comment was not deleted") unless destroy_comment
-    end
-
-    private
-
-    def destroy_comment
-      @destroy_comment = comment.destroy
+      context.fail!(error: "Comment was not deleted") unless comment.destroy
     end
   end
 end
