--- app/interactors/tasks/delete.rb
@@ -1,8 +1,0 @@
-module Tasks
-  class Delete
-    include Interactor::Organizer
-
-    organize Tasks::Destroy,
-             Tasks::Destroy::SendNotification
-  end
-end
