--- app/interactors/tasks/edit.rb
@@ -1,8 +1,0 @@
-module Tasks
-  class Edit
-    include Interactor::Organizer
-
-    organize Tasks::Update,
-             Tasks::Update::SendNotification
-  end
-end
