+++ app/interactors/tasks/update.rb
@@ -1,15 +1,8 @@
 module Tasks
   class Update
-    include Interactor
+    include Interactor::Organizer
 
-    delegate :task_params, :task, to: :context
-
-    def call
-      context.fail!(error: "something went wrong while updating task") unless update_task
-    end
-
-    def update_task
-      task.update(task_params)
-    end
+    organize Tasks::Update::Execute,
+             Tasks::Update::SendNotifications
   end
 end
