+++ config/database.yml
@@ -22,13 +22,8 @@ default: &default
   pool: <%= ENV.fetch("RAILS_MAX_THREADS") { 5 } %>
 
 development:
-  adapter: postgresql
-  encoding: unicode
-  database: MyDatabase
-  host: localhost
-  pool: 5
-  username: postgres
-  password: MyPassword
+  <<: *default
+  database: task_tracker_development
 
   # The specified database role being used to connect to postgres.
   # To create additional roles in postgres see `$ createuser --help`.
